Refactor/header

+++ src/redux/sagas/userSaga.ts
@@ -1,7 +1,7 @@
 import { fetchedData } from '@/components/Layouts/InfiniteScrollLayout';
 import { PayloadAction } from '@reduxjs/toolkit';
 import { call, put, takeLatest } from 'redux-saga/effects';
-import { authFail, authSuccess, reissueToken } from 'src/redux/reducers/auth';
+import { authFail, authSuccess } from 'src/redux/reducers/auth';
 import { openModal } from 'src/redux/reducers/components/modals';
 import {
   deleteUser,
@@ -44,29 +44,24 @@ import { patchPasswordType } from './../../services/UserService';
 function* getUserInfoSaga() {
   try {
     yield put(userPendingUserInfo());
-    const exp = TokenService.getExp();
-    if (exp && +exp - Math.floor(new Date().getTime() / 1000) < 30) {
-      const access = TokenService.get();
-      const refresh = TokenService.getRefresh();
-      yield put(reissueToken({ access, refresh }));
-    }
     const token = TokenService.get();
     appApi.defaults.headers.common['Authorization'] = `Bearer ${token}`;
     const userInfo: UserInfoType = yield call(UserService.getUserInfo);
     yield put(authSuccess(token));
     yield put(userSuccessUserInfo(userInfo));
   } catch (error: any) {
-    yield put(
-      userFailUserInfo(
-        new Error(error?.response?.data?.error || 'UNKNOWN_ERROR')
-      )
-    );
-    TokenService.remove();
-    TokenService.removeRefresh();
-    TokenService.removeExp();
-    yield put(
-      authFail(new Error(error?.response?.data?.error || 'UNKNOWN_ERROR'))
-    );
+    if (error?.response?.data?.error?.code !== 'DESTROY_JWT_TOKEN_EXCEPTION') {
+      yield put(
+        userFailUserInfo(
+          new Error(error?.response?.data?.error || 'UNKNOWN_ERROR')
+        )
+      );
+      TokenService.remove();
+      TokenService.removeExp();
+      yield put(
+        authFail(new Error(error?.response?.data?.error || 'UNKNOWN_ERROR'))
+      );
+    }
   }
 }
